test(handlers): break out handlers into modules for maintainability
