+++ lib/mocks/handlers/index.js
@@ -1,0 +1,29 @@
+const { rest } = require('msw');
+const upload = require('./upload');
+const authorize = require('./authorize');
+const image = require('./image');
+const gallery = require('./gallery');
+
+const handlers = [
+  //upload
+  rest.post('https://api.imgur.com/3/upload', upload.postHandler),
+
+  // gallery
+  rest.get('https://api.imgur.com/3/gallery/:id', gallery.getHandler),
+
+  // image
+  rest.post('https://api.imgur.com/3/image/:id', image.postHandler),
+  rest.post(
+    'https://api.imgur.com/3/image/:id/favorite',
+    image.postFavoriteHandler
+  ),
+  rest.delete('https://api.imgur.com/3/image/:id', image.deleteHandler),
+
+  // authorize
+  rest.get('https://api.imgur.com/oauth2/authorize', authorize.getHandler),
+  rest.post('https://api.imgur.com/oauth2/authorize', authorize.postHandler),
+];
+
+module.exports = {
+  handlers,
+};
